requirements.txt
flask
werkzeug
pillow
numpy
opencv-python
torch
torchvision
efficientnet_pytorch
onnxruntime
inference_sdk